SOF-1539 Refactor createStopAudioAction to be in Tv2AudioFactory
